Azure Function Token/requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
azure-identity
python-dotenv
PyJwt
azure-data-tables
azure-core
pytz
azure-keyvault-secrets
azure-keyvault-keys